Fix compilation buffer error scrolling
For some reason the logic to scroll to first error after
compilation failure would open the compilation buffer in the
current window (even if another window already had it open).
I put in a fix, with the caveat that nothing happens if no
window is showing the compilation buffer then nothing happens.